+++ src/system.c
@@ -37,7 +37,7 @@ void sys_start_frame()
     input_poll_and_load();
 }
 
-uint32_t *sys_get_frame_buffer() {
+uint8_t *sys_get_frame_buffer() {
     return ppu_get_frame_buffer();
 }
